Fix missing connection callbacks, Refactor CallbackInterface
The onConnected and onDisconnected callbacks for publishers, which are
suposed to fire if a listener connects or disconnects to the publisher,
were not working. It turned out that PeerConnDisconnCallback which
is inheriting from CallbackInterface was supposed to override a method
'Call', which it didn't and created a new method 'call'. To prevent further
cases of this CallbackInterface was turned into an abstract class,
improving the code quality while doing so.